third_party/WebKit/LayoutTests/fast/dom/custom/leaks-expected.txt

Tests that Custom Elements does not induce leaks

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test a custom element with minimal capturing
PASS observations.frame.wasCollected is true
PASS observations.window.wasCollected is true
PASS observations.document.wasCollected is true
PASS observations.proto.wasCollected is true
PASS observations.ctor.wasCollected is true
PASS observations.callback.wasCollected is true
Test callbacks which close over a bunch of objects
PASS observations.frame.wasCollected is true
PASS observations.window.wasCollected is true
PASS observations.document.wasCollected is true
PASS observations.proto.wasCollected is true
PASS observations.ctor.wasCollected is true
PASS observations.callback.wasCollected is true
Test that navigating a frame with custom elements does not leak
PASS observations.document.wasCollected is true
PASS observations.proto.wasCollected is true
PASS observations.ctor.wasCollected is true
PASS observations.callback.wasCollected is true
PASS successfullyParsed is true

TEST COMPLETE
